ordering_service/views.py
from rest_framework import viewsets
from django_filters.rest_framework import DjangoFilterBackend

from .serializers import OrderSerializer
from .models import Order


class OrderViewSet(viewsets.ModelViewSet):
    queryset = Order.objects.all().order_by('-created_at')
    serializer_class = OrderSerializer
    filter_backends = (DjangoFilterBackend,)
    filter_fields = ('status', 'customer_id')
