requirements.txt
Django==2.1.5
django-filter==2.0.0
djangorestframework==3.9.0
postgres==2.2.2
